Fix tests and eliminate the Box dtype warning
