+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import tempfile
 
 import pytest
+from gymnasium import register, registry
 from pytest import MonkeyPatch
 
 
@@ -12,3 +13,29 @@ def tmp_dataset_dir():
         mp.setenv("MINARI_DATASETS_PATH", tmp_dir.name)
         yield tmp_dir.name
     tmp_dir.cleanup()
+
+
+@pytest.fixture
+def register_dummy_envs():
+    env_names = [
+        "DummyBoxEnv",
+        "DummyInfoEnv",
+        "DummyInconsistentInfoEnv",
+        "DummyMultiDimensionalBoxEnv",
+        "DummyTupleDiscreteBoxEnv",
+        "DummyDictEnv",
+        "DummyTupleEnv",
+        "DummyTextEnv",
+        "DummyComboEnv",
+    ]
+    for env_name in env_names:
+        register(
+            id=f"{env_name}-v0",
+            entry_point=f"tests.common:{env_name}",
+            max_episode_steps=5,
+        )
+
+    yield
+
+    for env_name in env_names:
+        registry.pop(f"{env_name}-v0")
